cluster: async load rules if pd is api mode (#454)
* async load rules if pd is api mode


* region label don't expose api if not ready


* split and merge check labeler is ready


* add test for labler is ready


* add more test


* pause


* add more test


* pass ut


* address comment


---------

+++ pkg/schedule/labeler/labeler.go
@@ -17,8 +17,10 @@ package labeler
 import (
 	"context"
 	"strings"
+	"sync/atomic"
 	"time"
 
+	"github.com/pingcap/failpoint"
 	"github.com/pingcap/log"
 	"github.com/tikv/pd/pkg/core"
 	"github.com/tikv/pd/pkg/errs"
@@ -38,10 +40,11 @@ type RegionLabeler struct {
 	rangeList  rangelist.List // sorted LabelRules of the type `KeyRange`
 	ctx        context.Context
 	minExpire  *time.Time
+	isReady    atomic.Bool
 }
 
 // NewRegionLabeler creates a Labeler instance.
-func NewRegionLabeler(ctx context.Context, storage endpoint.RuleStorage, gcInterval time.Duration) (*RegionLabeler, error) {
+func NewRegionLabeler(ctx context.Context, storage endpoint.RuleStorage, gcInterval time.Duration, asyncLoad ...bool) (*RegionLabeler, error) {
 	start := time.Now()
 	defer func() {
 		newRegionLabelerDuration.Observe(time.Since(start).Seconds())
@@ -54,12 +57,44 @@ func NewRegionLabeler(ctx context.Context, storage endpoint.RuleStorage, gcInter
 		minExpire:  nil,
 	}
 
-	if err := l.loadRules(); err != nil {
-		return nil, err
+	if len(asyncLoad) > 0 && asyncLoad[0] {
+		go func() {
+			failpoint.Inject("pauseLoadRules", func(val failpoint.Value) {
+				if s, ok := val.(string); ok {
+					log.Info("skip loading region labeler rules", zap.String("duration", s))
+					expected, err := time.ParseDuration(s)
+					if err != nil {
+						log.Error("failed to parse duration", zap.String("duration", s), zap.Error(err))
+						return
+					}
+					select {
+					case <-time.After(expected):
+					case <-ctx.Done():
+						return
+					}
+				}
+			})
+			if err := l.loadRules(); err != nil {
+				log.Error("load rules failed", zap.Error(err))
+			}
+			log.Info("new region labeler created", zap.Int("label-rules-count", len(l.labelRules)))
+			go l.doGC(gcInterval)
+			l.isReady.Store(true)
+		}()
+	} else {
+		if err := l.loadRules(); err != nil {
+			return nil, err
+		}
+		l.isReady.Store(true)
+		log.Info("new region labeler created", zap.Int("label-rules-count", len(l.labelRules)))
+		go l.doGC(gcInterval)
 	}
-	log.Info("new region labeler created", zap.Int("label-rules-count", len(l.labelRules)))
-	go l.doGC(gcInterval)
 	return l, nil
+}
+
+// IsReady returns whether the labeler has loaded all rules.
+func (l *RegionLabeler) IsReady() bool {
+	return l.isReady.Load()
 }
 
 func (l *RegionLabeler) doGC(gcInterval time.Duration) {
@@ -113,6 +148,7 @@ func (l *RegionLabeler) checkAndClearExpiredLabels() {
 
 func (l *RegionLabeler) loadRules() error {
 	var toDelete []string
+	start := time.Now()
 	err := l.storage.LoadRegionRules(func(k, v string) {
 		r, err := NewLabelRuleFromJSON([]byte(v))
 		if err != nil {
@@ -128,6 +164,7 @@ func (l *RegionLabeler) loadRules() error {
 		}
 		l.labelRules[r.ID] = r
 	})
+	log.Info("load all region rules from storage", zap.Float64("take", time.Since(start).Seconds()))
 	if err != nil {
 		return err
 	}
@@ -139,6 +176,7 @@ func (l *RegionLabeler) loadRules() error {
 		}
 	}
 	l.BuildRangeListLocked()
+	log.Info("load region rules finished", zap.Float64("take", time.Since(start).Seconds()))
 	return nil
 }
 
@@ -162,6 +200,9 @@ func (l *RegionLabeler) BuildRangeListLocked() {
 
 // GetSplitKeys returns all split keys in the range (start, end).
 func (l *RegionLabeler) GetSplitKeys(start, end []byte) [][]byte {
+	if !l.IsReady() {
+		return nil
+	}
 	l.RLock()
 	defer l.RUnlock()
 	return l.rangeList.GetSplitKeys(start, end)
@@ -169,6 +210,9 @@ func (l *RegionLabeler) GetSplitKeys(start, end []byte) [][]byte {
 
 // GetAllLabelRules returns all the rules.
 func (l *RegionLabeler) GetAllLabelRules() []*LabelRule {
+	if !l.IsReady() {
+		return nil
+	}
 	l.checkAndClearExpiredLabels()
 	l.RLock()
 	defer l.RUnlock()
@@ -181,6 +225,9 @@ func (l *RegionLabeler) GetAllLabelRules() []*LabelRule {
 
 // GetLabelRules returns the rules that match the given ids.
 func (l *RegionLabeler) GetLabelRules(ids []string) ([]*LabelRule, error) {
+	if !l.IsReady() {
+		return nil, errs.ErrRuleNotFound.FastGenByArgs("region labeler is not ready")
+	}
 	now := time.Now()
 	rules := make([]*LabelRule, 0, len(ids))
 	for _, id := range ids {
@@ -193,6 +240,9 @@ func (l *RegionLabeler) GetLabelRules(ids []string) ([]*LabelRule, error) {
 
 // GetLabelRule returns the Rule with the same ID.
 func (l *RegionLabeler) GetLabelRule(id string) *LabelRule {
+	if !l.IsReady() {
+		return nil
+	}
 	return l.getAndCheckRule(id, time.Now())
 }
 
@@ -229,6 +279,9 @@ func (l *RegionLabeler) SetLabelRule(rule *LabelRule) error {
 
 // SetLabelRuleLocked inserts or updates a LabelRule but not buildRangeList.
 func (l *RegionLabeler) SetLabelRuleLocked(rule *LabelRule) error {
+	if !l.IsReady() {
+		return errs.ErrRuleNotFound.FastGenByArgs("region labeler is not ready")
+	}
 	if err := rule.checkAndAdjust(); err != nil {
 		return err
 	}
@@ -242,6 +295,9 @@ func (l *RegionLabeler) SetLabelRuleLocked(rule *LabelRule) error {
 // SaveLabelRuleLocked inserts or updates a LabelRule but not buildRangeList.
 // It only saves the rule to storage, and does not update the in-memory states.
 func (l *RegionLabeler) SaveLabelRuleLocked(rule *LabelRule) error {
+	if !l.IsReady() {
+		return errs.ErrRuleNotFound.FastGenByArgs("region labeler is not ready")
+	}
 	return l.storage.RunInTxn(l.ctx, func(txn kv.Txn) error {
 		return l.storage.SaveRegionRule(txn, rule.ID, rule)
 	})
@@ -249,6 +305,9 @@ func (l *RegionLabeler) SaveLabelRuleLocked(rule *LabelRule) error {
 
 // DeleteLabelRule removes a LabelRule.
 func (l *RegionLabeler) DeleteLabelRule(id string) error {
+	if !l.IsReady() {
+		return errs.ErrRuleNotFound.FastGenByArgs("region labeler is not ready")
+	}
 	l.Lock()
 	defer l.Unlock()
 	if _, ok := l.labelRules[id]; !ok {
@@ -263,6 +322,9 @@ func (l *RegionLabeler) DeleteLabelRule(id string) error {
 
 // DeleteLabelRuleLocked removes a LabelRule but not buildRangeList.
 func (l *RegionLabeler) DeleteLabelRuleLocked(id string) error {
+	if !l.IsReady() {
+		return errs.ErrRuleNotFound.FastGenByArgs("region labeler is not ready")
+	}
 	if err := l.storage.RunInTxn(l.ctx, func(txn kv.Txn) error {
 		return l.storage.DeleteRegionRule(txn, id)
 	}); err != nil {
@@ -274,6 +336,9 @@ func (l *RegionLabeler) DeleteLabelRuleLocked(id string) error {
 
 // Patch updates multiple region rules in a batch.
 func (l *RegionLabeler) Patch(patch LabelRulePatch) error {
+	if !l.IsReady() {
+		return errs.ErrRuleNotFound.FastGenByArgs("region labeler is not ready")
+	}
 	// setRulesMap is used to solve duplicate entries in DeleteRules and SetRules.
 	// Note: We maintain compatibility with the previous behavior, which is to process DeleteRules before SetRules
 	// If there are duplicate rules, we will prioritize SetRules and select the last one from SetRules.
@@ -324,6 +389,9 @@ func (l *RegionLabeler) Patch(patch LabelRulePatch) error {
 // GetRegionLabel returns the label of the region for a key.
 // If there are multiple rules that match the key, the one with max rule index will be returned.
 func (l *RegionLabeler) GetRegionLabel(region *core.RegionInfo, key string) string {
+	if !l.IsReady() {
+		return ""
+	}
 	l.RLock()
 	defer l.RUnlock()
 	now := time.Now()
@@ -350,6 +418,9 @@ func (l *RegionLabeler) GetRegionLabel(region *core.RegionInfo, key string) stri
 
 // ScheduleDisabled returns true if the region is lablelld with schedule-disabled.
 func (l *RegionLabeler) ScheduleDisabled(region *core.RegionInfo) bool {
+	if !l.IsReady() {
+		return true
+	}
 	v := l.GetRegionLabel(region, scheduleOptionLabel)
 	return strings.EqualFold(v, scheduleOptionValueDeny)
 }
@@ -357,6 +428,9 @@ func (l *RegionLabeler) ScheduleDisabled(region *core.RegionInfo) bool {
 // GetRegionLabels returns the labels of the region.
 // For each key, the label with max rule index will be returned.
 func (l *RegionLabeler) GetRegionLabels(region *core.RegionInfo) []*RegionLabel {
+	if !l.IsReady() {
+		return nil
+	}
 	l.RLock()
 	defer l.RUnlock()
 	type valueIndex struct {
